fix: code review core module
