+++ core/src/main/java/io/substrait/extendedexpression/ExtendedExpressionProtoConverter.java
@@ -1,4 +1,4 @@
-package io.substrait.extended.expression;
+package io.substrait.extendedexpression;
 
 import io.substrait.expression.Expression;
 import io.substrait.expression.proto.ExpressionProtoConverter;
@@ -7,13 +7,10 @@ import io.substrait.proto.ExpressionReference;
 import io.substrait.proto.ExtendedExpression;
 import io.substrait.type.proto.TypeProtoConverter;
 
-/**
- * Converts from {@link io.substrait.extended.expression.ExtendedExpression} to {@link
- * ExtendedExpression}
- */
+/** Converts from {@link ExtendedExpression} to {@link ExtendedExpression} */
 public class ExtendedExpressionProtoConverter {
   public ExtendedExpression toProto(
-      io.substrait.extended.expression.ExtendedExpression extendedExpression) {
+      io.substrait.extendedexpression.ExtendedExpression extendedExpression) {
 
     ExtendedExpression.Builder builder = ExtendedExpression.newBuilder();
     ExtensionCollector functionCollector = new ExtensionCollector();
@@ -21,30 +18,29 @@ public class ExtendedExpressionProtoConverter {
     final ExpressionProtoConverter expressionProtoConverter =
         new ExpressionProtoConverter(functionCollector, null);
 
-    for (io.substrait.extended.expression.ExtendedExpression.ExpressionReference
-        expressionReference : extendedExpression.getReferredExpr()) {
+    for (io.substrait.extendedexpression.ExtendedExpression.ExpressionReference
+        expressionReference : extendedExpression.getReferredExpressions()) {
 
       io.substrait.proto.Expression expressionProto =
           expressionProtoConverter.visit(
-              (Expression.ScalarFunctionInvocation) expressionReference.getReferredExpr());
+              (Expression.ScalarFunctionInvocation) expressionReference.getExpression());
 
       ExpressionReference.Builder expressionReferenceBuilder =
           ExpressionReference.newBuilder()
               .setExpression(expressionProto)
               .addAllOutputNames(expressionReference.getOutputNames());
 
-      extendedExpressionBuilder.addReferredExpr(expressionReferenceBuilder);
+      builder.addReferredExpr(expressionReferenceBuilder);
     }
-    extendedExpressionBuilder.setBaseSchema(
+    builder.setBaseSchema(
         extendedExpression.getBaseSchema().toProto(new TypeProtoConverter(functionCollector)));
 
     // the process of adding simple extensions, such as extensionURIs and extensions, is handled on
     // the fly
-    functionCollector.addExtensionsToExtendedExpression(extendedExpressionBuilder);
+    functionCollector.addExtensionsToExtendedExpression(builder);
     if (extendedExpression.getAdvancedExtension().isPresent()) {
-      extendedExpressionBuilder.setAdvancedExtensions(
-          extendedExpression.getAdvancedExtension().get());
+      builder.setAdvancedExtensions(extendedExpression.getAdvancedExtension().get());
     }
-    return extendedExpressionBuilder.build();
+    return builder.build();
   }
 }
