+++ core/src/main/java/io/substrait/plan/ProtoPlanConverter.java
@@ -32,10 +32,7 @@ public class ProtoPlanConverter {
   }
 
   public Plan from(io.substrait.proto.Plan plan) {
-    ExtensionLookup functionLookup =
-        ImmutableExtensionLookup.builder()
-            .from(plan.getExtensionUrisList(), plan.getExtensionsList())
-            .build();
+    ExtensionLookup functionLookup = ImmutableExtensionLookup.builder().from(plan).build();
     ProtoRelConverter relConverter = getProtoRelConverter(functionLookup);
     List<Plan.Root> roots = new ArrayList<>();
     for (PlanRel planRel : plan.getRelationsList()) {
