+++ core/src/test/java/io/substrait/extendedexpression/ProtoExtendedExpressionConverterTest.java
@@ -1,23 +1,24 @@
-package io.substrait.extended.expression;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
+package io.substrait.extendedexpression;
 
 import io.substrait.TestBase;
 import io.substrait.expression.Expression;
 import io.substrait.expression.ExpressionCreator;
 import io.substrait.expression.FieldReference;
 import io.substrait.expression.ImmutableFieldReference;
+import io.substrait.proto.ExtendedExpression;
 import io.substrait.type.ImmutableNamedStruct;
 import io.substrait.type.Type;
 import io.substrait.type.TypeCreator;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class ExtendedExpressionProtoConverterTest extends TestBase {
+public class ProtoExtendedExpressionConverterTest extends TestBase {
   @Test
-  public void toProtoTest() {
+  public void fromTest() throws IOException {
     // create predefined POJO extended expression
     Optional<Expression.ScalarFunctionInvocation> scalarFunctionExpression =
         defaultExtensionCollection.scalarFunctions().stream()
@@ -36,11 +37,12 @@ public class ExtendedExpressionProtoConverterTest extends TestBase {
 
     ImmutableExpressionReference expressionReference =
         ImmutableExpressionReference.builder()
-            .referredExpr(scalarFunctionExpression.get())
+            .expression(scalarFunctionExpression.get())
             .addOutputNames("new-column")
             .build();
 
-    List<ExtendedExpression.ExpressionReference> expressionReferences = new ArrayList<>();
+    List<io.substrait.extendedexpression.ExtendedExpression.ExpressionReference>
+        expressionReferences = new ArrayList<>();
     expressionReferences.add(expressionReference);
 
     ImmutableNamedStruct namedStruct =
@@ -50,24 +52,29 @@ public class ExtendedExpressionProtoConverterTest extends TestBase {
                 Type.Struct.builder()
                     .nullable(false)
                     .addFields(
-                        TypeCreator.NULLABLE.decimal(10, 2),
+                        TypeCreator.REQUIRED.decimal(10, 2),
                         TypeCreator.REQUIRED.STRING,
                         TypeCreator.REQUIRED.decimal(10, 2),
                         TypeCreator.REQUIRED.STRING)
                     .build())
             .build();
 
-    ImmutableExtendedExpression.Builder extendedExpression =
+    // pojo initial extended expression
+    ImmutableExtendedExpression extendedExpressionPojoInitial =
         ImmutableExtendedExpression.builder()
-            .referredExpr(expressionReferences)
-            .baseSchema(namedStruct);
+            .referredExpressions(expressionReferences)
+            .baseSchema(namedStruct)
+            .build();
 
-    // convert POJO extended expression into PROTOBUF extended expression
-    io.substrait.proto.ExtendedExpression proto =
-        new ExtendedExpressionProtoConverter().toProto(extendedExpression.build());
+    // proto extended expression
+    ExtendedExpression extendedExpressionProto =
+        new ExtendedExpressionProtoConverter().toProto(extendedExpressionPojoInitial);
 
-    assertEquals(
-        "/functions_arithmetic_decimal.yaml", proto.getExtensionUrisList().get(0).getUri());
-    assertEquals("add:dec_dec", proto.getExtensionsList().get(0).getExtensionFunction().getName());
+    // pojo final extended expression
+    io.substrait.extendedexpression.ExtendedExpression extendedExpressionPojoFinal =
+        new ProtoExtendedExpressionConverter().from(extendedExpressionProto);
+
+    // validate extended expression pojo initial equals to final roundtrip
+    Assertions.assertEquals(extendedExpressionPojoInitial, extendedExpressionPojoFinal);
   }
 }
